requirements.txt
PyYAML
jsonschema
requests
jsonpath_rw
ruamel.yaml
jsonschema2md
pandas
tabulate

